pyproject.toml
[tool.poetry]
name = "vitrina"
version = "0.1.0"
description = ""
authors = ["IVPK <atviriduomenys@ivpk.lt>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.13"
django-cms = "^3.10.0"
django-sekizai = "^3.0.1"
django-filer = "^2.2.1"
djangocms-text-ckeditor = "^5.0.1"
djangocms-blog = "^1.2.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"
WebTest = "^3.0.0"
django-webtest = "^1.9.10"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
